termwiz: fix panic when appending long runs to clusterline
There were two different overflows that could happen here when
resizing the terminal from a narrow width to a much wider width,
when the input consists of very very long logical lines.

The first was that we could overflow the maximum number of
cells in the ClusteredLine struct, which had u16 as its type.

The second, after changing u16 to u32, was the maximum length
of a run of cells with the same attribute.

The former could be fixed, preserving the smaller integer type,
by adjusting the API to allow it to return the overflow, but
that feels awkward and we could just increase the size of
the integer type.

The latter can be caught with an overflowing_add operation
and breaking the run in the overflow case.

That is what this commit does.

+bump codec version because this causes a small change
in the serialization of line data used by the mux
protocol.

refs: https://github.com/wezterm/wezterm/issues/6729